feat(Add postRequest middleware to options): Allow to pass an array of functions to run after the re
Adds a options.postRequest options to the initialization options that allows to pass an array of
functions that recieve the response as argument and should return the same response. This allows to
intercept all requests in a single place and act accordingly. We are assuming here, as with sdktor
in general, that the API that will be used is uniform across all endpoints.

Removes raw from the initialization options (it can be done in a postRequest middleware now)